Update crawl.py - Code Refactoring and Class distribution
Separated functions in to independent classes based on the service stack like all compute engine related functions under ComputeResourceMeta. This is just a proposal and will require making change in scanner.py for object declaration and method call
