+++ src/gcp_scanner/crawlers/__init__.py
@@ -1,0 +1,10 @@
+from .compute_crawler import ComputeResourceMeta
+from .db_crawler import DBResourceMeta
+from .gcs_crawler import GCSResourceMeta
+from .gke_crawler import GKEResourceMeta
+from .mq_crawler import MQResourceMeta
+from .network_crawler import NetworkResourceMeta
+from .project_crawler import ProjectResourceMeta
+from .repo_crawler import RepoResourceMeta
+from .security_crawler import SecurityResourceMeta
+from .serverless_crawler import ServerlessResourceMeta
