+++ src/gcp_scanner/crawlers/serverless_crawler.py
@@ -1,0 +1,75 @@
+import logging
+import sys
+from typing import List, Dict, Any
+
+from googleapiclient import discovery
+from httplib2 import Credentials
+
+class ServerlessResourceMeta:
+  def get_cloudfunctions(self,project_id: str,
+                        credentials: Credentials) -> List[Dict[str, Any]]:
+    """Retrieve a list of CloudFunctions available in the project.
+
+    Args:
+      project_id: A name of a project to query info about.
+      credentials: An google.oauth2.credentials.Credentials object.
+
+    Returns:
+      A list of CloudFunctions in the project.
+    """
+
+    logging.info("Retrieving CloudFunctions")
+    functions_list = list()
+    service = discovery.build(
+        "cloudfunctions", "v1", credentials=credentials, cache_discovery=False)
+    try:
+      request = service.projects().locations().functions().list(
+          parent=f"projects/{project_id}/locations/-")
+      while request is not None:
+        response = request.execute()
+        functions_list = response.get("functions", [])
+        request = service.projects().locations().functions().list_next(
+            previous_request=request, previous_response=response)
+    except Exception:
+      logging.info("Failed to retrieve CloudFunctions for project %s", project_id)
+      logging.info(sys.exc_info())
+
+    return functions_list
+
+  def get_app_services(self,project_name: str,
+                      credentials: Credentials) -> Dict[str, Any]:
+    """Retrieve a list of AppEngine instances available in the project.
+
+    Args:
+      project_name: A name of a project to query info about.
+      credentials: An google.oauth2.credentials.Credentials object.
+
+    Returns:
+      A dict representing default apps and services available in the project.
+    """
+
+    app_client = discovery.build(
+        "appengine", "v1", credentials=credentials, cache_discovery=False)
+
+    logging.info("Retrieving app services")
+    app_services = dict()
+    try:
+      request = app_client.apps().get(appsId=project_name)
+      response = request.execute()
+      if response.get("name", None) is not None:
+        app_services["default_app"] = (response["name"],
+                                      response["defaultHostname"],
+                                      response["servingStatus"])
+
+      request = app_client.apps().services().list(appsId=project_name)
+
+      app_services["services"] = list()
+      while request is not None:
+        response = request.execute()
+        app_services["services"] = response.get("services", [])
+        request = app_client.apps().services().list_next(
+            previous_request=request, previous_response=response)
+    except Exception:
+      logging.info("Failed to retrieve App services for project %s", project_name)
+      logging.info(sys.exc_info())
+    return app_services
